--- src/context.ts
@@ -1,3 +1,0 @@
-const context = globalThis as unknown as { useRequests: unknown };
-
-export default context;
