+++ src/global/context.ts
@@ -1,0 +1,5 @@
+import type { Context } from "@/types";
+
+const context = globalThis as unknown as Context;
+
+export default () => context.useRequests;
